+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 - Fixes for conflicting keybindings.
 - Windows installer is now signed with the Mapbox cert.
 - Windows installer now bundles Visual C++ (2013) runtimes.
+- Style UI: Display user modifications to layer list if present.
 
 ### 0.0.5